Editing functions simulating w/ Compadre data
This is a mess, and a bit of an incomplete commit. I've fixed a bunch
of issues with both the compadre simulation functions and also my
original simulation functions because they were not working properly
when using matrices with different dimensions (i.e., different om).
This is not yet complete. However, chatting with Lotte yesterday I
have also realized that I need to fix my extinction criteria. Blarg.