New titration funcs to find extinction thresh.
This commit adds new titration functions that explore sf x sm
parameter space and identify the extinction threshold. This is
a first pass, however. They are simple functions that click along
a sequence of sf values, titrating for the threshold sm value, and
then along sm values, titrating for the threshold sf value... but
this method can only works when there is a single threshold to find.
I have an idea to improve...