Fixed Bug in how lambda_i were calculated
This commit fixes an error in how lambda_i were being calculated
in several of the simulation functions. Basically, we were not
properly weighting the survival matrices (USi and UXi) when
calculating the genotype-specific growth rates, although we were
correctly weighting the Fertility matrices. This was causing our
titration functions to find incorrect invasion thresholds when
USi != UXi. The bug was most noticable when C = 0 and d_i > 0.