Add alpha argument to wrapper functions
This commit adds the density dependent growth shape parameter alpha
to all wrapper functions that make calls to the workhorse fwdDyn2Eq()
function. Default for all functions is for alpha = 0, which corresponds
to density-independent growth, as we had in the original simulations.
This *should* avoid any breakages due to adding in density dependent
growth.