Added .csv files to data folder
This commit adds .csv files containing the parameter values used in
the Mimulus case study as described in the main text and Appendix D.
The data files include:

1. data/peterson_2016_Data.csv
2. data/Willis_1993_Data.csv
3. data/inv6_Figure_Measurements.csv

(1) provides the demographic rates taken directly from Peterson et
al. (2016) Table 1. (see Corrected Table 1 in Peterson et al. 2017).

(2) Provides the data for selfing rate and inbreeding depression
terms taken from Tables 1 and 2 of Willis (1993).

(3) Provides the measurement data for figs 2,5,6 in Lee et al. (2016)
which is back-calculated to the original units and used to calculate
the selection coefficients on inv6 in:
./R/inv6-selection-coefficients.R