Use data files in run-Simulations.R
This commit alters the run-Simulations.R script so that prior to
running any simulations, it imports the .csv data files for the
Peterson et al. (2016) and Willis (1993) parameters reported in the
main text and Appendix D, and stores them as list objects. The lists
are then used for all simulations in the Mimulus case study.